Fixed: First pass at neume notation
The : character in the clef is now used to indicate neume notation, replacing the old '7.'

There may be other places in the guidelines that need fixing based on this change, but before doing too much it would be good to get this in place.

In light of this change, I have also clarified the meaning of the clefs with respect to octave, since neume clefs do not necessarily imply the octave in the same way. That probably needs to be worked through a bit more, but I thought it was a good idea to put the current practice in writing.

Fixes #4