Chords.
This is a first attempt at reaching some agreement on the encoding of chords.

Chords are enclosed in ^ and >. The duration of the chord is given prior to the opening ^ (similar to tuplets). No duration values are allowed within the chord. (i.e., chords with notes with differing durations are not allowed.)

Similar to repeat groups and repeat measures, tied chords are repeated with the v character. This means that you don't have to spell out the notes of the chords, and only identical tied chords are supported, varying only by their duration.

A p may be used to indicate a fermata over the chord.

Fixes #65
Fixes #60
Fixes #2