Fixed: Remove key and mode
The status of the "Key / Mode" field is ambiguous. It wasn't actually defined in Version 1, although the MARC21 documentation would seem to indicate otherwise.

Given the wide range of possible tonalities, particularly if we expand this to older music, Greek church modes, Indian music, etc. we can simply not carry this forward. To that end, a clarification in the Changes documentation is made.

Fixes #102